requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
pyusb==1.0.0rc1
requests==2.6.2
Werkzeug==0.10.4
wsgiref==0.1.2
